It works!
Big ol update to get the thing working
Can now modify scrap values and load the config properly

+++ WrenchOMatic/WrenchOMatic/Models/PlayerSettings.cs
@@ -8,12 +8,12 @@ namespace WrenchOMatic.Models
 {
     public class PlayerSettings
     {
-        public double ScrapMultiplier;
+        public int ScrapMultiplier;
         public bool AlwaysBigBonus;
-        public double ScrewSpeedMultiplier;
-        public double InspectMultiplier;
-        public double WalkingSpeedMultiplier;
-        public double PartMovementMultiplier;
+        public int ScrewSpeedMultiplier;
+        public int InspectMultiplier;
+        public int WalkingSpeedMultiplier;
+        public int PartMovementMultiplier;
 
         public PlayerSettings()
         {
